+++ ml/train/optimizers/optimizers.go
@@ -19,7 +19,6 @@
 package optimizers
 
 import (
-	"fmt"
 	. "github.com/gomlx/exceptions"
 	. "github.com/gomlx/gomlx/graph"
 	"github.com/gomlx/gomlx/ml/context"
@@ -137,11 +136,11 @@ func ByName(ctx *context.Context, optName string) Interface {
 	return optBuilder(ctx)
 }
 
-// GetGlobalStepVar returns the global step counter.
+// GetGlobalStepVar returns the global step counter, a dtypes.Int64 variable.
 // It creates it (initialized with 0) if not already there.
 // This can be used in graph building or directly.
 func GetGlobalStepVar(ctx *context.Context) *context.Variable {
-	return ctx.Checked(false).VariableWithValue(GlobalStepVariableName, 0).SetTrainable(false)
+	return ctx.Checked(false).VariableWithValue(GlobalStepVariableName, int64(0)).SetTrainable(false)
 }
 
 // GetGlobalStep returns the current global step value.
@@ -263,7 +262,6 @@ func addGradientsToVariablesGraph(ctx *context.Context, loss, learningRate, glob
 	}
 	numTrainable := len(grads)
 	ii := 0
-	fmt.Println("addGradientsToVariablesGraph")
 	ctx.EnumerateVariables(func(v *context.Variable) {
 		if !v.Trainable || !v.InUseByGraph(g) {
 			// Not interested in this variable.